+++ src/forward-projection.cpp
@@ -1,12 +1,12 @@
 #include "matrix.h"
 
-void forwardProjection(int** p, int rows, int cols) {
-  int** result = initializeMatrix(rows, cols);
+void forwardProjection(double** p, int rows, int cols) {
+  double** result = initializeMatrix(rows, cols);
   fillMatrix(result, rows, cols, 0);
 
   // rotating the image from 0 to 179 degrees and accumulating the pixel values in the result matrix
   for (int i = 0; i < 180; i++) {
-    int** temp = rotateMatrix(p, rows, cols, (-1) * i);
+    double** temp = rotateMatrix(p, rows, cols, (-1) * i);
 
     /**
      * Accumulating stage
